+++ CHANGES.rst
@@ -1,6 +1,12 @@
 version 0.8.1dev (unreleased)
 -----------------------------
-
+- code cleanup and minor bug fixes
+- background fit added to 1D and 2D Gaussian fits
+- plotting AiryDisk2D fit is now possible
+- unit test updates
+- new options added to the aperture phot parameter set that allow users to plot the used apertures
+- ZScaleInterval added from astropy.virtualization to set the color range on the data for aperture photometry plot
+- replaced the sigma to fwhm lambda with the astropy constant for conversion
 
 version 0.8.0 (released)
 ------------------------